#39 💷 refactor : 메인페이지 관련 컴포넌트 리팩토링
