+++ .cspell.json
@@ -1,0 +1,3 @@
+{
+  "words": ["tanstack", "firestore", "Galmuri"]
+}
